New class for the graph, handles errors differently
I'm still working on Gio's last upload to see if this is something we want.  The only thing this implementation doesn't do is check for cycles.  That feature doesn't work here.  Take a look; but this is only part of what I've been working on.  Hopefully this is going in the right direction.